requirement.txt
flask
psycopg2
bcrypt
python-dotenv